Use package module to install distro packages
Consolidate distro package install tasks into a
single task using the package module. Tidy up
some other tasks to reduce task file sprawl and
consolidate some task actions.

The minimum Ansible version is raised to 2.2 due to a
known bug [1] in Ansible's apt module which does not
update the cache properly if the cache update and the
install are combined in a single task.

[1] https://github.com/ansible/ansible-modules-core/issues/1497

Change-Id: I9752dffdb7302d2d76f2cf0c7023cc3535f3fc9f