Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: If553ad33bd071792198239bc567c2e06058ebddd
Partial-Bug: #1764470
(cherry picked from commit 23b68a02919b114c261dfa2571c01e33ee30dc3e)