Set quota for octavia service project
The default quotas are often not sufficient for a side wide
load balancing service. This will adjust them to some more
reasonable values.

The os_quota module of ansible assumes an installed
block storage service so it is not suitable for our
tests which don't install cinder. Therefore we
are using the openstack client straight.

Change-Id: If85cc3862b9316d9bf43e025bfaab8e1b7fd5576