Clean up test override vars
Several override vars aren't used, or are only used in one
test playbook will will get revamped to eliminate the need
for them after the service delegation patch merges.

As such, we remove the unnecessary vars and move those to
be cleaned up later into the play to make it simpler to do.

Change-Id: Idfcf5fb400079eadd080b30992a90cc758e1cac8