Execute image setup against octavia_service_setup_host
In order to reduce the packages required to pip install on to the hosts,
we use service delegation to octavia_service_setup_host so that instead
of installing software on the target host, and putting credentials on
every target host, we isolate the software and credentials to a single
host.

In this patch we remove the variable 'octavia_image_downloader' and replace
it with just using the 'octavia_service_setup_host' instead. We also need
to add the variable 'octavia_amp_image_path_owner' which is set to the user
running the playbook by default, so that the image can be downloaded to the
deployment host successfully.

There are any other tasks in the role which need updating before we can
eliminate the octavia_requires_pip_packages, but for the sake of keeping
the patch smaller and easier to review they will be done in follow up
patches.

Change-Id: I438cdf695abe223a9fcf7ead796fe2eef41845b7