Execute security group setup against octavia_service_setup_host
In order to reduce the packages required to pip install on to the hosts,
we use service delegation to octavia_service_setup_host so that instead
of installing software on the target host, and putting credentials on
every target host, we isolate the software and credentials to a single
host.

In this patch we make the tasks execute on octavia_service_setup_host so
that we do not need as much software installed on the target host. We
retain the use of the 'auth' arguments because the security group must
be created in the octavia service project, and the os_security_group
module does not support allowing an admin to create them in a different
project.

There are any other tasks in the role which need updating before we can
eliminate the octavia_requires_pip_packages, but for the sake of keeping
the patch smaller and easier to review they will be done in follow up
patches.

Change-Id: Idc8e5aa922541ab5f8464c83c006b1b7fc8308c5