Remove octavia_requires_pip_packages
In order to reduce the packages required to pip install on to the hosts,
we use service delegation to octavia_service_setup_host so that instead
of installing software on the target host, and putting credentials on
every target host, we isolate the software and credentials to a single
host.

In this patch we finally remove octavia_requires_pip_packages as it is
no longer required given that all services use delegation to the
octavia_service_setup_host now. We also remove the task which used to
install them.

We also remove the meta dependency on the openstack_openrc role because
the target host does not require openrc/clouds.yaml any more.

We remove the variable 'octavia_ansible_endpoint_type' as it is no longer
used.

We finally remove the python-pip package from the distro packages, as
with these changes, pip is no longer required on the host - everything
installed by pip is isolated into a venv.

Change-Id: If3f65fdb8a59a3c41cb8c2f0dee6b2e5a71f05c7