Ensure the octavia flavor is private
In https://review.openstack.org/588166 we switch to
using clouds.yaml, but in doing so made the flavor
creation happen in the admin project instead of the
octavia project.

This patch ensures that the flavor is created in the
right project.

We also do away with the extra task to get the flavor
ID because the task to create it will return results
including the ID whether it has created a new one or
it has found an existing one.

Change-Id: I24e89a0466c8b0cef632f1677809e04622f22786