Clean up role tests
Now that we're using the general templates, we can slim down
the role test definitions. We can also remove tests which are
not being watched, or which are fundamentally broken. With
this we can also remove unused scripts/plays.

We do the following:

1. We remove the 'ssl' job, given that the person working on
   that is no longer doing so, and no-one else has picked up
   the work.

Change-Id: I2ce20adf933e050e7429d6876fa1f0049db4183d
(cherry picked from commit 873b7585a54fd1805ab44cbb12efc83bc7739052)