Only run OpenStack tasks once
By having run_once on the entire include, the set_fact only
runs on a single node which means the configuration for all
the other nodes fails.

This patch makes the play get included for everything, but
only the OpenStack creation happens once, with fact setting
happening on all nodes.

Change-Id: Iddd4afebb491dcf9f78f13b7ad96574f8a68e074