Stop arbitrary shutting down services
It looks like services are shut down for no reason if we're running
a migration which can mean they will never be started back up if
a playbook fails.

This patch stops shutting down services for no reason as most likely
the service will be restarted after the packages change.

Change-Id: If4214abf928711c364d7962a0369db4a69a4440f