Ensure that the flavor/network facts are available to all targets
Currently the registered result for the flavor creation is
only available on localhost. To ensure that it is available
on all hosts we first create it using the designated service
host, then fetch the details and set a fact for all hosts.

We do the same for the management network fact.

In order to achieve this, shade needs to be present on the
target host, so we add it into the octavia venv. We also make
sure that the octavia service account is used, because that's
where the flavor and network objects were created and only that
account has access to them.

Change-Id: I85b533dbb5fa29382dc615b584b21ff1a5d67a46