Fix updating amphora images
The octavia_amp_image.yml playbook is not idempotent because it is
missing some conditional logic and uses an incorrect result variable
name for the `os_image_facts` module (albeit following the module
documentation).

This change also updates the use of the `id` key to the required
`name` key in the `os_image` module.

Also, the fact get and image deletion need to be done by the service
account, not the cloud admin account.

Finally, these tasks all have the same conditional, so we put them
all into a block.

Change-Id: I52e111fffe9a60c879c967c94a895ae272d98fcd