Enable overriding the service setup host python interpreter
In order to enable the service setup host python interpreter to
be changed easily, we make it a variable. This will be useful
when someone sets the service setup host to be the utility
container, because we'll be able to set this var by default.

Change-Id: I3ad7e3006180b0790232b0a6ccd4b57e5be7a30d
(cherry picked from commit c265ae146bf7e7f10c76ddd53472574f51056b86)