Test using integrated build
In order to do a more complete verification of any patches,
we add a full uncontainerised OpenStack deployment to do the
functional testing using the integrated repo. This replaces
the previous functional test mechanism.

Any additional role tests are left as-is. They will require
some extra implementation in the integrated build before they
can be transferred.

This also includes a fix in the role which starts using the
os_quota role as the version of Ansible we ship has resolved
that bug.

In addition, the container network name added a trailing
addres which caused the lookup to fail.  This resolves that
issue so the configuration file can be dropped cleanly.

Depends-On: https://review.openstack.org/647840
Depends-On: https://review.openstack.org/648502
Depends-On: https://review.openstack.org/648551
Depends-On: https://review.openstack.org/648579
Depends-On: https://review.openstack.org/648793
Change-Id: I689c679c91c5d516cb8285077c7d3c7c064c9112