keypair: copy key to deploy host rather than setup host
The env lookup in the key copying task refers to the deploy host.
If your deploy host user is not root then this task will fail
as this directory does not exist on the remote setup host.

As the original intention appears to have been to copy the result
of the keypair generation to the deploy host, the delegation is
explicitly set to 'localhost'.

Change-Id: I89649503d5918c33f0d1e4200be67be5e0ed8a9e
(cherry picked from commit ffac17201a735115267648b657399ad9fc4be9ae)