Set CentOS 9 Stream as NV
Currently CentOS 9 Stream hits libvirt bug [1] that occurs on attaching
interface to running VM. Octavia is most affected by that as this is
part of usual workflow.
Fix for the bug has been already proposed and should be released
quite soon. Not to spend a lot of time on that, we rather wait for new
libvirt version to be released  for CentOS 9 Stream and until then
set jobs as NV.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2092856

Change-Id: I9dd6fcea23154f781ec111a0927a26aba28954f1