Ensure quota is set for BFV amphoras
At the moment role does not increase volumes quota for the project
when BFV is enabled for Octavia. This might lead to Amphora
failures to spawn-up once default quota (of 10 volumes) is reached.

With that we also calculate amout of diskspace required based on the
instance limit.

It also changes flavor to contain 0 disk if BFV is enabled to ensure
that ephemeral is not gonna be used.

Change-Id: Ia35948cd2381b13d3792fc0029d93a9e176c884c